+++ migrations-legacy/20211231000001-add-last_deployed-to-guardian.js
@@ -4,8 +4,8 @@ module.exports = {
   up: function (queryInterface, Sequelize) {
     return queryInterface.sequelize.transaction(async t => {
       await queryInterface.sequelize.query(`
-        IF NOT EXISTS( 
-          SELECT NULL FROM information_schema.COLUMNS 
+        IF NOT EXISTS(
+          SELECT NULL FROM information_schema.COLUMNS
           WHERE table_name = 'Guardians' AND column_name = 'last_deployed' AND table_schema = 'rfcx_api'
         ) THEN
           ALTER TABLE Guardians ADD last_deployed datetime(3) NULL AFTER stream_id;