+++ tasks/event-names.js
@@ -1,4 +1,5 @@
 const SEGMENT_CREATED = 'segmentCreated'
 const EVENT_CREATED = 'event-created'
+const EVENT_UPDATED = 'event-updated'
 
-module.exports = { SEGMENT_CREATED, EVENT_CREATED }
+module.exports = { SEGMENT_CREATED, EVENT_CREATED, EVENT_UPDATED }